app/requirements.txt
Flask==2.0.3
Flask-RESTful==0.3.9
flask-swagger==0.2.14
gunicorn==20.1.0
Flask-Cors == 3.0.10
Werkzeug == 2.0.3
pymongo == 4.0.2
Flask-PyMongo == 2.3.0
dnspython == 2.2.1